requirements.txt
pdfplumber==0.11.4
pandas==2.2.2
rapidfuzz==3.8.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0